+++ lib/global_id/locator.rb
@@ -22,8 +22,8 @@ class GlobalID
       # The GlobalIDs are located using Model.find(array_of_ids), so the models must respond to
       # that finder signature.
       #
-      # This approach will efficiently call only one #find per model class, but still interpolate
-      # the results to match the order in which the gids were passed.
+      # This approach will efficiently call only one #find (or #where(id: id), when using ignore_missing)
+      # per model class, but still interpolate the results to match the order in which the gids were passed.
       #
       # Options:
       # * <tt>:only</tt> - A class, module or Array of classes and/or modules that are
@@ -31,11 +31,17 @@ class GlobalID
       #   classes to those classes or their subclasses.  Passing one or more modules in limits
       #   instances of returned classes to those including that module.  If no classes or
       #   modules match, +nil+ is returned.
+      # * <tt>:ignore_missing</tt> - By default, locate_many will call #find on the model to locate the
+      #   ids extracted from the GIDs. In Active Record (and other data stores following the same pattern),
+      #   #find will raise an exception if a named ID can't be found. When you set this option to true,
+      #   we will use #where(id: ids) instead, which does not raise on missing records.
       def locate_many(gids, options = {})
         if (allowed_gids = parse_allowed(gids, options[:only])).any?
           models_and_ids  = allowed_gids.collect { |gid| [ gid.model_name.constantize, gid.model_id ] }
           ids_by_model    = models_and_ids.group_by(&:first)
-          loaded_by_model = Hash[ids_by_model.map { |model, ids| [ model, model.find(ids.map(&:last)).index_by { |record| record.id.to_s } ] }]
+          loaded_by_model = Hash[ids_by_model.map { |model, ids|
+            [ model, find_records(model, ids.map(&:last), ignore_missing: options[:ignore_missing]).index_by { |record| record.id.to_s } ]
+          }]
 
           models_and_ids.collect { |(model, id)| loaded_by_model[model][id] }.compact
         else
@@ -115,6 +121,14 @@ class GlobalID
 
         def normalize_app(app)
           app.to_s.downcase
+        end
+
+        def find_records(model_class, ids, ignore_missing:)
+          if ignore_missing
+            model_class.where(id: ids)
+          else
+            model_class.find(ids)
+          end
         end
     end