+++ test/models/person.rb
@@ -1,16 +1,27 @@
 class Person
   include GlobalID::Identification
 
+  HARDCODED_ID_FOR_MISSING_PERSON = '1000'
+
   attr_reader :id
 
   def self.find(id_or_ids)
     if id_or_ids.is_a? Array
       ids = id_or_ids
-      ids.collect { |id| new(id) }
+      ids.collect { |id| find(id) }
     else
       id = id_or_ids
-      new(id)
+
+      if id == HARDCODED_ID_FOR_MISSING_PERSON
+        raise 'Person missing'
+      else
+        new(id)
+      end
     end
+  end
+
+  def self.where(conditions)
+    (conditions[:id] - [HARDCODED_ID_FOR_MISSING_PERSON]).collect { |id| new(id) }
   end
 
   def initialize(id = 1)